+++ alicloud/provider.go
@@ -97,6 +97,7 @@ func Provider() terraform.ResourceProvider {
 			"alicloud_slb_listeners":        dataSourceAlicloudSlbListeners(),
 			"alicloud_slb_rules":            dataSourceAlicloudSlbRules(),
 			"alicloud_slb_server_groups":    dataSourceAlicloudSlbServerGroups(),
+			"alicloud_oss_buckets":          dataSourceAlicloudOssBuckets(),
 			"alicloud_db_instances":         dataSourceAlicloudDBInstances(),
 			"alicloud_pvtz_zones":           dataSourceAlicloudPvtzZones(),
 			"alicloud_pvtz_zone_records":    dataSourceAlicloudPvtzZoneRecords(),